Se configura base de datos
