+++ src/anonimacion/api/__init__.py
@@ -2,21 +2,17 @@ import os
 
 from flask import Flask, jsonify, request
 from flask_swagger import swagger
-from salud_tech.modulos.procesamiento.aplicacion.servicios import ServicioDatasetMedico
-from salud_tech.modulos.procesamiento.aplicacion.dto import DatasetMedicoDto
-from salud_tech.modulos.procesamiento.dominio.entidades import DatasetMedico
-
 
 # Identifica el directorio base
 basedir = os.path.abspath(os.path.dirname(__file__))
 
 def registrar_handlers():
     # Importa los handlers
-    import salud_tech.modulos.procesamiento.aplicacion
+    import anonimacion.modulos.procesamiento.aplicacion
 
 def importar_modelos_alchemy():
     # Import SQLAlchemy models
-    import salud_tech.modulos.procesamiento.infraestructura.dto
+    import anonimacion.modulos.procesamiento.infraestructura.dto
 
 def comenzar_consumidor():
     """
@@ -26,7 +22,7 @@ def comenzar_consumidor():
     """
 
     import threading
-    import salud_tech.modulos.procesamiento.infraestructura.consumidores as procesamiento
+    import anonimacion.modulos.procesamiento.infraestructura.consumidores as procesamiento
 
     # Suscripción a eventos
     threading.Thread(target=procesamiento.suscribirse_a_eventos).start()
@@ -38,10 +34,10 @@ def create_app(configuracion={}):
     # Init la aplicacion de Flask
     app = Flask(__name__, instance_relative_config=True)
     
-    db_user = os.getenv('POSTGRES_USER', 'salud_tech')
-    db_password = os.getenv('POSTGRES_PASSWORD', 'salud_tech_123')
-    db_name = os.getenv('POSTGRES_DB', 'rabbit_salud_tech')
-    db_host = os.getenv('POSTGRES_HOST', 'salud_tech_db')
+    db_user = os.getenv('POSTGRES_USER', 'anonimacion')
+    db_password = os.getenv('POSTGRES_PASSWORD', 'anonimacion_123')
+    db_name = os.getenv('POSTGRES_DB', 'rabbit_anonimacion')
+    db_host = os.getenv('POSTGRES_HOST', 'anonimacion_db')
     
     app.config['SQLALCHEMY_DATABASE_URI'] =\
             f'postgresql://{db_user}:{db_password}@{db_host}/{db_name}'
@@ -51,7 +47,7 @@ def create_app(configuracion={}):
     app.config['TESTING'] = configuracion.get('TESTING')
 
      # Inicializa la DB
-    from salud_tech.config.db import init_db, db
+    from anonimacion.config.db import init_db, db
     init_db(app)
      
     with app.app_context():
@@ -59,21 +55,20 @@ def create_app(configuracion={}):
         if not app.config.get('TESTING'):
             comenzar_consumidor()
 
-    print(db, "db despues de inicializar db")
     importar_modelos_alchemy()
     registrar_handlers()
 
      # Importa Blueprints
-    from . import procesamiento
+    from . import anonimacion
 
     # Registro de Blueprints
-    app.register_blueprint(procesamiento.bp)
+    app.register_blueprint(anonimacion.bp)
 
     @app.route("/spec")
     def spec():
         swag = swagger(app)
         swag['info']['version'] = "1.0"
-        swag['info']['title'] = "Salud Tech - Procesamiento API"
+        swag['info']['title'] = "Salud Tech - Anonimacion API"
         return jsonify(swag)
 
     @app.route("/health")
