+++ src/anonimacion/modulos/anonimacion/aplicacion/queries/obtener_dataset_medico.py
