+++ src/anonimacion/modulos/anonimacion/dominio/entidades.py
