+++ src/anonimacion/modulos/anonimacion/dominio/reglas.py
