+++ src/anonimacion/modulos/anonimacion/infraestructura/consumidores.py
