+++ src/anonimacion/modulos/anonimacion/infraestructura/dto.py
