+++ src/anonimacion/modulos/anonimacion/infraestructura/schema/v1/eventos.py
